+++ src/egon/data/datasets/emobility/motorized_individual_travel/__init__.py
@@ -413,7 +413,7 @@ class MotorizedIndividualTravel(Dataset):
 
         super().__init__(
             name="MotorizedIndividualTravel",
-            version="0.0.1",
+            version="0.0.3",
             dependencies=dependencies,
             tasks=(
                 create_tables,