+++ src/egon/data/datasets/sanity_checks.py
@@ -12,12 +12,12 @@ class SanityChecks(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="SanityChecks",
-            version="0.0.2",
+            version="0.0.3",
             dependencies=dependencies,
-            tasks=(
+            tasks={
                 sanitycheck_eGon2035_electricity,
                 sanitycheck_eGon2035_heat,
-            ),
+            },
         )